pyproject.toml
[tool.poetry]
name = "song-summary"
version = "0.1.0"
description = "Test Task"
authors = ["Artur Tsarikovich"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
openai = "^1.55.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
